Bug 648776 - mirror symmetries.
You can now set any paint tool to mirror any stroke relatively
horizontal/vertical axis or a central point (any combination of these 3
symmetries).
This has been implemented as a new multi-stroke core, where every stroke
is actually handled as a multi-stroke (of course by default, the multi-stroke
is of size 1).
Current version has to be activated in the playground.